Add Mechanism For Omitting Content From Slides (#33)
Add mechanism for omitting content from slides

The slides in this project can be consumed in two ways:

1. As a slideshow presented to a group of people
2. As reading material for people to peruse in their own time

It's useful to be able to adapt the content to support these two scenarios. For example, having a lot of copy available to describe the concepts is useful in scenario 2), because there's no presenter available; however, having a large amount of copy in scenario 1) is not desirable, because it can distract from the presentation.

This change introduces a way to selectively hide content from the slide build pipeline, but keep it available when viewing the slides in raw markdown or when processed by standard markdown parsers (e.g. on GitHub).

It does this by relying on additional utilities from the `unifiedjs` ecosystem. Specifically, the `mdast-zone` utility is used to select "zones" of content within specific markdown comments, and remove these zones using a small custom plugin.

```
Content to remain in the slides

// custom-comment-start
Content to omit
// custom-comment-end

More content to remain in the slides
```